% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mode.R
\name{get_mode}
\alias{get_mode}
\title{Get the mode (most frequent) of a vector}
\usage{
get_mode(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a scalar
}
\description{
Get the mode (most frequent) of a vector
}
\examples{
get_mode(c(1, 2, 2, 3))

}
