% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_constant_columns.R
\name{is_constant}
\alias{is_constant}
\alias{is_not_constant}
\title{Check if the input is constant (excluding NA)}
\usage{
is_constant(var)

is_not_constant(var)
}
\arguments{
\item{var}{a vector, list, dataframe column}
}
\value{
TRUE or FALSE
}
\description{
Check if the input is constant (excluding NA)
}
\examples{
is_constant(1:10)
is_not_constant(1:10)


}
